chore(refactor): unflatten core services to allow for future services to join in the ranks
