+++ src/app/rules/consecutive/index.ts
@@ -4,7 +4,7 @@ import {
   GridNodeTraverser,
   Direction,
 } from '@app/models/grid-linked-list/grid-node-traverser';
-import { GridRoom } from '@app/core/room-repository.service';
+import { GridRoom } from '@app/core/room-repo/room-repository.service';
 
 export class ConsecutiveRule extends Rule {
   constructor(
